// src/app/models/symbol.model.ts
export interface Symbol {
  title?: string;
  symbol?: string;
  card?: string;
  image?: string;
  status?: boolean;
  receive?: boolean;
  send?: boolean;
  cardSpendable?: boolean;
  conversionRate?: any;
  order?: number;
  color?: string;
  fee?: number;
  feeFormatted?: string;
  marketData?: any;
  extra?: any;
  type?: string;
  id?: string;
}